researchers/analyzers/evm/requirements.txt
web3==6.11.3
requests==2.31.0
python-dotenv==1.0.0 